+++ packages/plugin-hotkey/src/const.ts
@@ -1,0 +1,46 @@
+export const HOTKEY_MAP = {
+  /**
+   * 历史记录回退
+   */
+  HISTORY_UNDO: ['command+z', 'ctrl+z'],
+
+  /**
+   * 历史记录前进
+   */
+  HISTORY_REDO: ['command+y', 'ctrl+y'],
+
+  /**
+   * 锁定/解锁
+   */
+  LOCK_UNLOCK: ['command+shift+l', 'ctrl+shift+l'],
+
+  /**
+   * 显示/隐藏
+   */
+  SHOW_HIDE: ['command+shift+h', 'ctrl+shift+h'],
+
+  /**
+   * 复制
+   */
+  COPY: ['command+c', 'ctrl+c'],
+
+  /**
+   * 粘贴
+   */
+  PASTE: ['command+v', 'ctrl+v'],
+
+  /**
+   * 剪切
+   */
+  CUT: ['command+x', 'ctrl+x'],
+
+  /**
+   * 删除
+   */
+  DELETE: ['backspace', 'del'],
+
+  /**
+   * 取消选中
+   */
+  CLEAR_SELECTION: ['esc'],
+}