+++ packages/core/CHANGELOG.md
@@ -8,6 +8,12 @@
 - 1c9ec92: chore: remove package's CHANGELOG.md
 - 6f282a3: build: move mobx,mobx-react to peerDependencies
 
+## 0.0.3-alpha.2
+
+### Patch Changes
+
+- bd41334: fix: document.import error
+
 ## 0.0.3-alpha.1
 
 ### Patch Changes