+++ packages/react-renderer/CHANGELOG.md
@@ -12,6 +12,13 @@
 - Updated dependencies [6f282a3]
   - @easy-editor/core@0.0.3
 
+## 0.0.2-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [bd41334]
+  - @easy-editor/core@0.0.3-alpha.2
+
 ## 0.0.2-alpha.1
 
 ### Patch Changes